Move utils functions
