+++ tests/openstack/test_images.py
@@ -8,12 +8,8 @@ from openstack.image.v2.member import Member
 from pytest_cases import case, parametrize, parametrize_with_cases
 
 from src.providers.openstack import get_images
-from tests.schemas.utils import (
-    random_image_os_type,
-    random_image_status,
-    random_image_visibility,
-    random_lower_string,
-)
+from tests.openstack.utils import random_image_status, random_image_visibility
+from tests.schemas.utils import random_image_os_type, random_lower_string
 
 
 # `shared` has a separate case
