+++ tests/openstack/test_networks.py
@@ -9,11 +9,8 @@ from pytest_cases import case, parametrize, parametrize_with_cases
 
 from src.models.provider import PrivateNetProxy
 from src.providers.openstack import get_networks
-from tests.schemas.utils import (
-    random_ip,
-    random_lower_string,
-    random_network_status,
-)
+from tests.openstack.utils import random_network_status
+from tests.schemas.utils import random_ip, random_lower_string
 
 
 @case(tags=["is_shared"])
